Add in useful helicopters and make piloting and mechanic free (#9)
Purpose of change
Very few Cataclysm players will ever even find a helicopter, much less have the abilities to pilot it. While it may not be "realistic", the gameplay of helicopters is so ridiculously gated that nobody can experience it, and the inability to modify helicopters removes the fun of being able to build your own vehicles for this unique vehicle type. Really anti-player design here.

Describe the solution
- Port in Useful Helicopters mod
- Make "Piloting" and "Airframe & Powerplant" mechanic free at chargen
    - Remove the skill bonuses these skills give so they don't give a real advantage

At a later date, I think it would make sense to outright strip the restrictrions from piloting and helicopter modification to make this more seamless. But that would take some C++ edits, and this is "good enough" for the time being.