Disable Portal Storms (#7)
Purpose of change
Portal storms are cool in theory/lore, but are really non-interactive gameplay in practice. Either you're inside and you're safe, or your outside and you're dead. Add to that the fact that portal storms will fuck with things like animals, the inclusion of portal storms incentivises players to do weird stuff like store all their cattle in the basement.

This feature is not cooked for regular play. Let's disable it. If we ever make them better, we might re-enable.

Describe the solution
Remove EOC that triggers portal storms