Remove AFS Formless Ruins (#17)
Summary
Content "Remove AFS Formless Ruins"

Purpose of change
The Aftershock Formless Ruins are one of those bits of content that snuck into Aftershock while seemingly being intended for Exoplanet. They stick out like a sore thumb and are disruptive of intended thematics in Cataclysm 2040.

Describe the solution
I hate removing content (especially in a fork about content restoration), but these ruins are really designed to represent space habitats and don't transfer over well to the 2040s content that Aftershock originally was a dumping ground for.

Describe alternatives you've considered
I thought about reworking the designs and item groups, but it would just take a lot of time. On top of the fact that my time is limited and I simply don't have time for it at the moment, I think that time would be better spent building original 2040 POIs rather than trying to rework this content that was never meant to be representative of the original 2040 setpieces in C:DDA.

Testing
Added this to a save and revealed overmap to ensure these aren't being spawned anymore. I even added it to an existing save to make sure it got rid of the ruins properly.